old.requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
ci-info==0.2.0
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.1
cymem==2.0.6
docx==0.2.4
docx2pdf==0.1.8
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
etelemetry==0.3.0
filelock==3.6.0
fitz==0.0.1.dev2
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.2.0/fr_core_news_md-3.2.0-py3-none-any.whl
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.2.0/fr_core_news_sm-3.2.0-py3-none-any.whl
future==0.18.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
lxml==4.7.1
MarkupSafe==2.0.1
murmurhash==1.0.6
mysql-connector==2.2.9
networkx==2.6.3
nibabel==3.2.2
nipype==1.7.0
nltk==3.6.7
pandas==1.4.1
pathlib==1.0.1
pathy==0.6.1
pdf2image==1.16.0
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.0
preshed==3.0.6
prov==2.0.0
pycparser==2.21
pydantic==1.8.2
pydot==1.4.2
PyMuPDF==1.19.5
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pyxnat==1.4
rdflib==6.1.1
regex==2022.1.18
requests==2.27.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
starlette==0.18.0
thinc==8.0.13
tqdm==4.62.3
traits==6.3.2
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.5
Wand==0.6.7
wasabi==0.9.0